% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialsource-package.R
\docType{package}
\name{specialsource-package}
\alias{specialsource}
\alias{specialsource-package}
\title{specialsource: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
