% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_code.R
\name{tag-code}
\alias{tag-code}
\title{Include a function's source code in its documentation}
\description{
Use this tag when documenting a function to insert that function's source
code into the documentation itself.
}
\keyword{internal}
